tarot backend

+++ src/__main__.py
@@ -1,20 +1,22 @@
 import json
+import os
 
 from bson import ObjectId
 from bson.json_util import dumps
-from flask import Flask, json, jsonify, request
+from flask import Flask, jsonify, request
 from flask_cors import CORS
-from flask_jwt import JWT, jwt_required, current_identity
+from flask_jwt_extended import create_access_token, get_jwt_identity, jwt_required, JWTManager
+from werkzeug.security import generate_password_hash, check_password_hash
 
 from src import db, env
-from src.auth import authenticate, identity
 
 app = Flask(__name__)
 CORS(app)
 
 app.secret_key = env.SECRET_KEY
-
-jwt = JWT(app, authenticate, identity)
+jwt = JWTManager(app)
+db_username = os.getenv('UPORABNIK')
+db_geslo = os.getenv('GESLO')
 
 
 @app.route('/api/blog', methods=['GET'])
@@ -29,51 +31,79 @@ def get_blog_id(_id):
     return json.loads(objava)
 
 
+@app.route("/api/login", methods=["POST"])
+def login():
+    username = request.json.get("username")
+    password = request.json.get("password")
+
+    if not username or not password:
+        return jsonify({"msg": "Missing username or password"}), 400
+
+    if username != db_username or password != db_geslo:
+        return jsonify({"msg": "Bad username or password"}), 401
+
+    access_token = create_access_token(identity=username)
+    return jsonify(access_token=access_token)
+
+
+@app.route('/api/blog_pregled', methods=['GET'])
+@jwt_required()
+def get_blog_pregled():
+    current_user = get_jwt_identity()
+    if current_user:
+        return json.loads(dumps(db.proces.objava.find()))
+    return jsonify(logged_in_as=current_user), 200
+
+
 @app.route('/api/blog', methods=['POST'])
 @jwt_required()
 def add_blog():
-    print(current_identity)
+    current_user = get_jwt_identity()
     data = request.get_json()
     if data is not None:
         db.proces.objava.insert_one(data)
         return jsonify({"message": "Objava uspešno dodana!"})
 
-    return jsonify({"error": "Neveljavna zahteva ali manjkajoči podatki!"})
+    return jsonify(logged_in_as=current_user), 200
 
 
 @app.route('/api/admin')
 @jwt_required()
 def get_admin():
-    print(current_identity)
-    return jsonify({"message": "Ste na admin strani!"})
+    current_user = get_jwt_identity()
+    return jsonify(logged_in_as=current_user), 200
 
 
 @app.route('/api/blog/delete/<_id>', methods=['DELETE'])
 @jwt_required()
 def delete_blog(_id):
-    print(current_identity)
-    if _id is not None:
-        db.proces.objava.delete_one({'_id': ObjectId(_id)})
-        return jsonify({"message": "Blog izbrisanm uspešno"})
-    else:
-        return jsonify({"error": "Blog ne obstaja"})
+    current_user = get_jwt_identity()
+    if current_user:
+        if _id is not None:
+            db.proces.objava.delete_one({'_id': ObjectId(_id)})
+            return jsonify({"message": "Blog izbrisanm uspešno"})
+        else:
+            return jsonify({"error": "Blog ne obstaja"})
+    return jsonify(logged_in_as=current_user), 200
 
 
 @app.route('/api/blog/edit/<_id>', methods=['POST'])
 @jwt_required()
 def update_blog(_id):
-    print(current_identity)
-    data = request.get_json()
-    result = db.proces.objava.update_one({'_id': ObjectId(_id)}, {'$set': data})
-    if result.modified_count > 0:
-        updated_document = db.proces.objava.find_one({'_id': ObjectId(_id)})
-        updated_document['_id'] = str(updated_document['_id'])
-        return jsonify({"message": "Objava uspešno posodobljena", "updated_document": updated_document})
-    else:
-        return jsonify({"error": "Objava bloga ni uspela!"})
+    current_user = get_jwt_identity()
+    if current_user:
+        data = request.get_json()
+        result = db.proces.objava.update_one({'_id': ObjectId(_id)}, {'$set': data})
+        if result.modified_count > 0:
+            updated_document = db.proces.objava.find_one({'_id': ObjectId(_id)})
+            updated_document['_id'] = str(updated_document['_id'])
+            return jsonify({"message": "Objava uspešno posodobljena", "updated_document": updated_document})
+        else:
+            return jsonify({"error": "Objava bloga ni uspela!"})
+    return jsonify(logged_in_as=current_user), 200
 
 
 if __name__ == '__main__':
-    db.drop()
-    db.seed()
+    # db.drop()
+    # db.seed()
     app.run(host='0.0.0.0', port=env.PORT, debug=True)
